Implement adapter model for tracked signals
The previous method of tracking signals was to store refernces to them
in tracked_signals. To add signals to master_output, these references
were gotten from tracked_signals and put in master_output's
input_signals. Since tracked_signals is manipulated regularly, and the
signal references within it changed and freed, this method was not
dynamic, and highly unstable. To work, it would have required
notifying all its users after every modification to keep their
references up to date. One solution would have been for the mixer to
store GtkTreePaths to signals in tracked_signals and reference them
for every audio sample, but that would have been comparatively slow
and required the Signal framework to be dependent on GTK. Instead,
signals are simply chained to an adapter, and a pointer to that
adapter is stored in tracked_signals. The adapter contains a pointer
to the signal, which can safely be changed. This way, a
DeluneSignalComponent can be set to control a parameter of another
component, and the controlling signal can be dynamically changed.